Moving stats methods to extra lib module
