--- src/view/empty-stats.js
@@ -1,5 +1,0 @@
-import Abstract from "./abstract.js";
-
-export default class EmptyStats extends Abstract {
-
-}
